Recomposite object layers onto video outputs. New salient object removal process: generate dirty run inpaint with high CFG for interrogation, then inpaint normally using auto-generated prompt. Handling wider range of layer numbers. set ffmpeg params

+++ src/layers/salient_object.py
@@ -1,15 +1,17 @@
 from moviepy.editor import VideoClip, ImageClip
+from moviepy.video.fx import mask_color
 from interfaces.project_interface import ProjectInterface
 from interfaces.layer_interface import LayerInterface
 from interfaces.logger_interface import LoggerInterface
 import os
+import numpy as np
+import cv2
 from PIL import Image
 from comfy_api.client import ComfyClient
 from comfy_api.server import ComfyServer
 from workflow_wrapper.workflow import ComfyAPIWorkflow
 from utils.update_path_parts import update_path_parts
 from termcolor import colored
-
 from constants import (
     SALIENT_OBJECTS_WORKFLOW_PATH,
     FEATHERING_MARGIN,
@@ -40,47 +42,53 @@ class SalientObjectLayer(LayerInterface):
         )
         self.__update_workflow()
 
-        # Construct comfy client with the custom workflow
-        try:
-            server = ComfyServer(
-                self.project,
-                self.logger,
-                output_directory=self.project.project_dir_path,
-                input_directory=self.project.project_dir_path,
-            )
-            server.start()
+        self.alpha_layer_fullpath, self.base_layer_fullpath = self.__get_comfy_outputs()
 
-            client = ComfyClient(self.project, self.workflow, self.logger)
-            client.queue_workflow()
-        except Exception as e:
-            self.logger.log(f"Error starting comfy server/client: {e}")
-            raise e
-        finally:
+        # Only generate salient objects layers and new base layer if they don't already exist
+        if (
+            not self.alpha_layer_fullpath
+            or input("\n\nUse existing salient object layers? (y/n):\n> ").lower()
+            != "y"
+        ):
+            # Construct comfy client with the custom workflow
             try:
-                server.kill()
-                client.disconnect()
+                server = ComfyServer(
+                    self.project,
+                    self.logger,
+                    output_directory=self.project.project_dir_path,
+                    input_directory=self.project.project_dir_path,
+                )
+                server.start()
+
+                client = ComfyClient(self.project, self.workflow, self.logger)
+                client.queue_workflow()
             except Exception as e:
-                self.logger.log(f"Error stopping comfy server/client: {e}")
+                self.logger.log(f"Error starting comfy server/client: {e}")
+                raise e
+            finally:
+                try:
+                    server.kill()
+                    client.disconnect()
+                except Exception as e:
+                    self.logger.log(f"Error stopping comfy server/client: {e}")
 
-        alpha_layer = [
-            f
-            for f in os.listdir(self.project.project_dir_path)
-            if SALIENT_OBJECT_ALPHA_LAYER_PREFIX in f
-        ][-1]
-        self.alpha_layer_fullpath = os.path.join(self.project.project_dir_path, alpha_layer)
-        base_layer = [
-            f
-            for f in os.listdir(self.project.project_dir_path)
-            if BASE_LAYER_WITHOUT_OBJECTS_PREFIX in f
-        ][-1]
+            self.alpha_layer_fullpath, self.base_layer_fullpath = (
+                self.__get_comfy_outputs()
+            )
 
+        self.logger.log(
+            f"Salient Object Alpha Layer: {self.alpha_layer_fullpath}",
+            pad_with_rules=False,
+        )
+        self.logger.log(
+            f"Base Layer Without Objects: {self.base_layer_fullpath}",
+            pad_with_rules=False,
+        )
         # Set the inpainted base image (with objects removed) as the new input image
         self.project.update_config(
             "input_image_original_path", self.project.config_file()["input_image_path"]
         )
-        self.project.update_config(
-            "input_image_path", os.path.join(self.project.project_dir_path, base_layer)
-        )
+        self.project.update_config("input_image_path", self.base_layer_fullpath)
 
         self.__set_original_layer()
         self.__set_layer_breakpoints()
@@ -112,8 +120,8 @@ class SalientObjectLayer(LayerInterface):
         )
         self.output_vid_width = self.original_layer["image"].width
 
-        # Create an empty alpha image with height the same as the base image and width = (total_steps * velocity) - ((total_steps + 1) * feathering_margin)
-        width = (self.total_steps * self.get_x_velocity()) - (
+        # Create an empty alpha image with height the same as the base image and width = original + slide distance
+        width = (self.total_steps * abs(self.get_x_velocity())) - (
             (self.total_steps + 1) * FEATHERING_MARGIN
         )
         alpha_image = Image.new(
@@ -157,8 +165,23 @@ class SalientObjectLayer(LayerInterface):
     def create_layer_videoclip(self) -> VideoClip:
         image_clip = ImageClip(self.stitched["fullpath"])
 
+        # Convert alpha channel to binary mask
+        cv2_image = cv2.imread(self.stitched["fullpath"], cv2.IMREAD_UNCHANGED)
+        height, width, _ = cv2_image.shape
+        alpha_channel = cv2_image[:, :, 3]
+        mask = np.zeros((height, width), dtype=np.uint8)
+        for i in range(height):
+            for j in range(width):
+                if alpha_channel[i, j] > 0:
+                    mask[i, j] = 1
+
+        def make_mask_frame(t):
+            # Calculate the position based on time
+            x = int(self.slide_distance * (t / self.duration))
+            # Take a cropping of the mask from x to x + self.output_vid_width
+            return mask[:, x : x + self.output_vid_width]
+
         def make_frame(t):
-            # Print progress and x-coordinate of the layer every 10 seconds if devmode
             if DEV and t % 10 == 0 and t != 0:
                 print(
                     colored(f"{self.name_prefix} at {t}seconds", "light_green"),
@@ -176,7 +199,28 @@ class SalientObjectLayer(LayerInterface):
             x = int(self.slide_distance * (t / self.duration))
             return image_clip.get_frame(t)[:, x : x + self.output_vid_width]
 
-        return VideoClip(make_frame, duration=self.duration)
+        mask_video = VideoClip(make_mask_frame, duration=self.duration, ismask=True)
+
+        ret = VideoClip(make_frame, duration=self.duration)
+        ret = ret.set_mask(mask_video)
+        return ret
+
+    def get_x_velocity(self):
+        # Add logic to clean, adjust, or change type of velocity
+        # NOTE: for now, make velocity of salient objects slightly slower to make them stand out
+        # return int(self.layer_config["velocity"][0] * 0.6)
+        return int(self.layer_config["velocity"][0])
+
+    def get_y_velocity(self):
+        # Add logic to clean, adjust, or change type of velocity
+        return int(self.layer_config["velocity"][1])
+
+    def get_final_layer_height(self):
+        # Add cleaning, adjusting, or type changing logic
+        return int(self.original_layer["image"].height)
+
+    def get_final_layer_width(self):
+        return self.slide_distance + self.original_layer["image"].width
 
     def __set_original_layer(self):
         self.original_layer = {
@@ -203,9 +247,7 @@ class SalientObjectLayer(LayerInterface):
     def __find_lowest_non_alpha_pixel(self):
         width, height = self.original_layer["image"].size
         lowest_pixel = None
-
         last_layer_breakpoint = self.layer_height_breakpoints[-1]
-
         for y in range(height - 1, -1, -1):
             for x in range(width):
                 pixel = self.original_layer["image"].getpixel((x, y))
@@ -218,37 +260,29 @@ class SalientObjectLayer(LayerInterface):
             if lowest_pixel:
                 break
 
+        return lowest_pixel
+
     def __set_parent_layer(self):
         """Determine the parent layer for this salient object.
         First, determine the lowest point in the salient object original image which has a non-alpha pixel (i.e., the lowest non-alpha pixel).
         Then, determine which layer contains that point.
         """
         lowest_non_alpha_pixel = self.__find_lowest_non_alpha_pixel()
-        self.logger.log(f"Lowest non-alpha pixel: {lowest_non_alpha_pixel}")
+        self.logger.log(
+            f"Lowest non-alpha pixel: {lowest_non_alpha_pixel}", pad_with_rules=False
+        )
 
         # Default is the last layer (lowest) because if the loop doesn't break, then the salient object's lowest non-alpha pixel isnt below any of the layer breakpoints
         self.parent_layer_index = len(self.layer_height_breakpoints) - 1
         for index, breakpoint in enumerate(self.layer_height_breakpoints):
-            self.logger.log(f"Testing layer {index+1} breakpoint: {breakpoint}")
+            self.logger.log(
+                f"Testing layer {index+1} breakpoint: {breakpoint}",
+                pad_with_rules=False,
+            )
             if lowest_non_alpha_pixel[1] < breakpoint:
                 self.parent_layer_index = index
                 self.logger.log(f"Salient object's lowest point is in layer {index+1}")
                 break
-
-    def get_x_velocity(self):
-        # Add logic to clean, adjust, or change type of velocity
-        return int(self.layer_config["velocity"][0])
-
-    def get_y_velocity(self):
-        # Add logic to clean, adjust, or change type of velocity
-        return int(self.layer_config["velocity"][1])
-
-    def get_final_layer_height(self):
-        # Add cleaning, adjusting, or type changing logic
-        return int(self.original_layer["image"].height)
-
-    def get_final_layer_width(self):
-        return self.slide_distance + self.original_layer["image"].width
 
     def __update_workflow(self):
         # Set the input image for the salient object workflow as the project's input image
@@ -283,7 +317,7 @@ class SalientObjectLayer(LayerInterface):
             self.prompt_tags.replace(" . ", ", "),
             append=True,
         )
-        
+
         # Also add to the auto-prompt exclude list so the tagger that generates the prompt doesn't include the salient object tags
         self.workflow.update(
             "Auto Prompt Exclude List",
@@ -300,3 +334,25 @@ class SalientObjectLayer(LayerInterface):
         )
 
         self.workflow.save()
+
+    def __get_comfy_outputs(self):
+        try:
+            alpha_layer = [
+                f
+                for f in os.listdir(self.project.project_dir_path)
+                if SALIENT_OBJECT_ALPHA_LAYER_PREFIX in f
+            ][-1]
+            base_layer = [
+                f
+                for f in os.listdir(self.project.project_dir_path)
+                if BASE_LAYER_WITHOUT_OBJECTS_PREFIX in f
+            ][-1]
+            return (
+                os.path.join(self.project.project_dir_path, alpha_layer),
+                os.path.join(self.project.project_dir_path, base_layer),
+            )
+        except IndexError:
+            self.logger.log(
+                "Salient Object Alpha Layers: Not found",
+            )
+            return (False, False)
